feat: separate scale into interface and implementation
This separation removes the requirement that scales are based on
arrays of intervals.
The IScale interface now abstractly defines what a scale is, allowing
us to implement scales based on sequences of notes, intervals or even
other mechanisms.
The ScaleWalker implements a scale degree based method for traversing
and by extension constructing scales, allowing us to construct most
musical scales.
The SymmtericScale constructs scales whose intervals are the same
regardless of the direction in which they are traversed.

This feature allows us to implement the melodic minor scale later on.